reorg 5b: move workspaces/ (obsidian) + control/cmux/ bridge into @cockpit/workspaces
