+++ packages/workspaces/src/index.ts
@@ -1,3 +1,7 @@
 // @cockpit/workspaces — environment/surface seam (cmux runtime · obsidian vault · cmux notifier).
 export * from "./runtimes/index.js";
 export * from "./notifiers/index.js";
+export * from "./workspaces/index.js";
+export { CmuxEventsBridge, deriveRunState } from "./cmux/events-bridge.js";
+export type { RunState, CmuxEventsChild, CmuxAgentHook, CmuxEventsBridgeDeps } from "./cmux/events-bridge.js";
+export { DaemonCmux } from "./cmux/daemon-cmux.js";
