+++ packages/workspaces/src/workspaces/__tests__/obsidian.test.ts
