+++ packages/workspaces/src/workspaces/__tests__/registry.test.ts
