+++ packages/workspaces/src/workspaces/obsidian.ts
