+++ src/commands/doctor.ts
@@ -7,8 +7,7 @@ import chalk from "chalk";
 import { loadConfig } from "@cockpit/shared";
 import { compatManifest, type ToolEntry } from "@cockpit/shared";
 import { checkToolCompat } from "@cockpit/shared";
-import { createCmuxDriver, RuntimeRegistry, createCmuxNotifier, NotifierRegistry } from "@cockpit/workspaces";
-import { createObsidianDriver, WorkspaceRegistry } from "../workspaces/index.js";
+import { createCmuxDriver, RuntimeRegistry, createCmuxNotifier, NotifierRegistry, createObsidianDriver, WorkspaceRegistry } from "@cockpit/workspaces";
 import { queryHealth, printServiceHealth } from "./health-view.js";
 import {
   createCursorEmitter,
