+++ src/commands/launch.ts
@@ -8,8 +8,7 @@ import chalk from "chalk";
 import { loadConfig, resolveHome, type ModelRoutingConfig } from "@cockpit/shared";
 import { createClaudeDriver, createCodexDriver, createGeminiDriver, createOpencodeDriver, CapabilityRegistry } from "@cockpit/agents";
 import type { AgentDriver, Role } from "@cockpit/agents";
-import { RuntimeRegistry, createCmuxDriver } from "@cockpit/workspaces";
-import { createObsidianDriver, WorkspaceRegistry } from "../workspaces/index.js";
+import { RuntimeRegistry, createCmuxDriver, createObsidianDriver, WorkspaceRegistry } from "@cockpit/workspaces";
 import type { RuntimeDriver } from "@cockpit/workspaces";
 import { ensureSpokeLayout } from "@cockpit/shared";
 import { resolveCmuxBin } from "@cockpit/shared";
