+++ src/commands/status.ts
@@ -2,7 +2,7 @@ import { Command } from "commander";
 import chalk from "chalk";
 import matter from "gray-matter";
 import { loadConfig } from "@cockpit/shared";
-import { createObsidianDriver, WorkspaceRegistry } from "../workspaces/index.js";
+import { createObsidianDriver, WorkspaceRegistry } from "@cockpit/workspaces";
 import { queryHealth, printServiceHealth } from "./health-view.js";
 
 interface StatusFrontmatter {
