+++ src/lib/__tests__/canonical-source.test.ts
@@ -3,7 +3,7 @@ import os from "node:os";
 import fsp from "node:fs/promises";
 import path from "node:path";
 import { readUserLevelSource, readProjectLevelSource } from "@cockpit/shared";
-import { createObsidianDriver } from "../../workspaces/index.js";
+import { createObsidianDriver } from "@cockpit/workspaces";
 import type { WorkspaceDriver } from "@cockpit/shared";
 
 function memDriver(files: Record<string, string>): WorkspaceDriver {
